[ubicloud] Rebuild apt lists from scratch
When we booted a vm from generated image, we found that apt-get update
failed with the following error previously:

    E: Encountered a section with no Package: header
    E: Problem with MergeList /var/lib/apt/lists/_etc_apt_apt-mirrors.txt_dists_jammy_multiverse_i18n_Translation-en
    E: The package lists or status file could not be parsed or opened.

In the previous cases, generating a new image was solving the issue. But
for the current version, I created a new image at least 10 times but the
issue was still there. So I decided to remove the apt lists and rebuild
them from scratch. It's happening for only Ubuntu 22.04 x64 image.