[ubicloud] Persist GitHub context to a file
GitHub environment variables are only available in the workflow, not for
unix user. We need some of these variables to run some features such as
Ubicloud Cache. Runner script allows to run a hook script before the job
starts and this hook has access to the environment variables. So we just
persist them to a file and read them in the Ubicloud Cache proxy.