Add a feature to display messages on the GitHub UI
GitHub Actions has workflow commands that let you send messages to the
GitHub UI. This can be helpful for giving feedback or information during
a workflow's execution to our customers. The control plane saves the
message to a file, and the start and complete hooks print them if the
files exist.

Since the file needs to be in workflow command format, GitHub Actions
automatically parses the output and shows the message in the UI.

https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/workflow-commands-for-github-actions#setting-a-notice-message